backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
groq==0.4.2
google-generativeai==0.3.2
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.26.0